requirements.txt
asgiref==3.3.4
attrs==20.3.0
awsebcli==3.19.4
bcrypt==3.2.0
blessed==1.18.0
botocore==1.20.53
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2
django-extensions==3.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
future==0.16.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
semantic-version==2.8.5
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.58.0
